+++ src/main/groovy/org/jenkinsci/gradle/plugins/jpi/JpiHplManifest.groovy
@@ -16,6 +16,7 @@
 package org.jenkinsci.gradle.plugins.jpi
 
 import org.gradle.api.Project
+import org.gradle.api.java.archives.Attributes
 import org.gradle.api.plugins.WarPlugin
 import org.gradle.api.plugins.WarPluginConvention
 import org.gradle.api.tasks.bundling.War
@@ -26,16 +27,23 @@ import org.gradle.api.tasks.bundling.War
 class JpiHplManifest extends JpiManifest {
     JpiHplManifest(Project project) {
         super(project)
+    }
+
+    @Override
+    Attributes getAttributes() {
+        Attributes mainAttributes = super.attributes
 
         def conv = project.extensions.getByType(JpiExtension)
         War war = project.tasks.getByName(WarPlugin.WAR_TASK_NAME) as War
 
         // src/main/webApp
         def warconv = project.convention.getPlugin(WarPluginConvention)
-        mainAttributes.putValue('Resource-Path', warconv.webAppDir.absolutePath)
+        mainAttributes['Resource-Path'] = warconv.webAppDir.absolutePath
 
         // add resource directories directly so that we can pick up the source, then add all the jars and class path
         Set<File> libraries = conv.mainSourceTree().resources.srcDirs + war.classpath.files
-        mainAttributes.putValue('Libraries', libraries.findAll { it.exists() }.join(','))
+        mainAttributes['Libraries'] = libraries.findAll { it.exists() }.join(',')
+
+        mainAttributes
     }
 }