+++ src/test/groovy/org/jenkinsci/gradle/plugins/jpi/JpiManifestSpec.groovy
@@ -43,7 +43,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('basics.mf')
@@ -69,6 +69,53 @@ version = '1.2'"""
         new JarFile(generatedJarFile).manifest.mainAttributes == readManifest('test.mf').mainAttributes
     }
 
+    def 'JAR is up-to-date if unchanged'() {
+        given:
+        temporaryFolder.newFolder('test', 'src', 'main', 'java')
+        temporaryFolder.newFile('test/build.gradle') << PROJECT
+        temporaryFolder.newFile('test/src/main/java/TestPlugin.java') << 'class TestPlugin extends hudson.Plugin {}'
+        GradleRunner runner = GradleRunner.create()
+                .withProjectDir(new File(temporaryFolder.root, 'test'))
+                .withPluginClasspath()
+                .withArguments('jar')
+
+        when:
+        BuildResult result = runner.build()
+
+        then:
+        result.task(':jar').outcome == TaskOutcome.SUCCESS
+
+        when:
+        result = runner.build()
+
+        then:
+        result.task(':jar').outcome == TaskOutcome.UP_TO_DATE
+    }
+
+    def 'JAR is built if manifest changed'() {
+        given:
+        temporaryFolder.newFolder('test', 'src', 'main', 'java')
+        File buildFile = temporaryFolder.newFile('test/build.gradle') << PROJECT
+        temporaryFolder.newFile('test/src/main/java/TestPlugin.java') << 'class TestPlugin extends hudson.Plugin {}'
+        GradleRunner runner = GradleRunner.create()
+                .withProjectDir(new File(temporaryFolder.root, 'test'))
+                .withPluginClasspath()
+                .withArguments('jar')
+
+        when:
+        BuildResult result = runner.build()
+
+        then:
+        result.task(':jar').outcome == TaskOutcome.SUCCESS
+
+        when:
+        buildFile.text = PROJECT.replace('1.2', '1.3')
+        result = runner.build()
+
+        then:
+        result.task(':jar').outcome == TaskOutcome.SUCCESS
+    }
+
     def 'JPI contains manifest'() {
         given:
         temporaryFolder.newFolder('test', 'src', 'main', 'java')
@@ -83,10 +130,62 @@ version = '1.2'"""
                 .build()
 
         then:
+        result.task(':war').outcome == TaskOutcome.SUCCESS
         result.task(':jpi').outcome == TaskOutcome.SUCCESS
         File generatedJarFile = new File(temporaryFolder.root, 'test/build/libs/test.hpi')
         generatedJarFile.exists()
         new JarFile(generatedJarFile).manifest.mainAttributes == readManifest('test.mf').mainAttributes
+    }
+
+    def 'JPI is up-to-date if unchanged'() {
+        given:
+        temporaryFolder.newFolder('test', 'src', 'main', 'java')
+        temporaryFolder.newFile('test/build.gradle') << PROJECT
+        temporaryFolder.newFile('test/src/main/java/TestPlugin.java') << 'class TestPlugin extends hudson.Plugin {}'
+        GradleRunner runner = GradleRunner.create()
+                .withProjectDir(new File(temporaryFolder.root, 'test'))
+                .withPluginClasspath()
+                .withArguments('jpi')
+
+        when:
+        BuildResult result = runner.build()
+
+        then:
+        result.task(':war').outcome == TaskOutcome.SUCCESS
+        result.task(':jpi').outcome == TaskOutcome.SUCCESS
+
+        when:
+        result = runner.build()
+
+        then:
+        result.task(':war').outcome == TaskOutcome.UP_TO_DATE
+        result.task(':jpi').outcome == TaskOutcome.UP_TO_DATE
+    }
+
+    def 'JPI is built if manifest changed'() {
+        given:
+        temporaryFolder.newFolder('test', 'src', 'main', 'java')
+        File buildFile = temporaryFolder.newFile('test/build.gradle') << PROJECT
+        temporaryFolder.newFile('test/src/main/java/TestPlugin.java') << 'class TestPlugin extends hudson.Plugin {}'
+        GradleRunner runner = GradleRunner.create()
+                .withProjectDir(new File(temporaryFolder.root, 'test'))
+                .withPluginClasspath()
+                .withArguments('jpi')
+
+        when:
+        BuildResult result = runner.build()
+
+        then:
+        result.task(':war').outcome == TaskOutcome.SUCCESS
+        result.task(':jpi').outcome == TaskOutcome.SUCCESS
+
+        when:
+        buildFile.text = PROJECT.replace('1.2', '1.3')
+        result = runner.build()
+
+        then:
+        result.task(':war').outcome == TaskOutcome.SUCCESS
+        result.task(':jpi').outcome == TaskOutcome.SUCCESS
     }
 
     def 'plugin class'() {
@@ -104,7 +203,7 @@ version = '1.2'"""
         new File(directory, 'hudson.Plugin').write('org.example.PluginImpl')
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('plugin-class.mf')
@@ -117,7 +216,7 @@ version = '1.2'"""
         }
 
         when:
-        JpiManifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest.mainAttributes.getValue('Plugin-Version') =~
@@ -140,7 +239,7 @@ version = '1.2'"""
         (project as ProjectInternal).evaluate()
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('dependency.mf')
@@ -163,7 +262,7 @@ version = '1.2'"""
         (project as ProjectInternal).evaluate()
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('dependencies.mf')
@@ -185,7 +284,7 @@ version = '1.2'"""
         (project as ProjectInternal).evaluate()
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('optional-dependency.mf')
@@ -208,7 +307,7 @@ version = '1.2'"""
         (project as ProjectInternal).evaluate()
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('optional-dependencies.mf')
@@ -233,7 +332,7 @@ version = '1.2'"""
         (project as ProjectInternal).evaluate()
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('complex-dependencies.mf')
@@ -252,7 +351,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('compatible-since-version.mf')
@@ -271,7 +370,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('mask-classes.mf')
@@ -290,7 +389,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('plugin-first-class-loader.mf')
@@ -309,7 +408,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('sandbox-status.mf')
@@ -334,7 +433,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('plugin-developer.mf')
@@ -362,7 +461,7 @@ version = '1.2'"""
         }
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest('plugin-developers.mf')
@@ -385,7 +484,7 @@ version = '1.2'"""
         new File(directory, 'hudson.Extension').bytes = index
 
         when:
-        Manifest manifest = new JpiManifest(project)
+        Manifest manifest = new JpiManifest(project).toJavaManifest()
 
         then:
         manifest == readManifest("support-dynamic-loading-${value}.mf")